Add recipe for conditionally typechecking gems
It took me a few tries to figure out the best way to include the stubs in a gem I'm developing.  My goal was to have strict typechecking during development, but no typechecking when imported into a larger project.  This is working nicely for me and thought I'd share since it's probably a pretty common use case.

(Thanks for all your work on sorbet tooling!)